+++ LibGit2Sharp.Tests/StageFixture.cs
@@ -21,7 +21,7 @@ namespace LibGit2Sharp.Tests
             string path = SandboxStandardTestRepo();
             using (var repo = new Repository(path))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
                 Assert.Equal(doesCurrentlyExistInTheIndex, (repo.Index[relativePath] != null));
                 Assert.Equal(currentStatus, repo.RetrieveStatus(relativePath));
 
@@ -39,7 +39,7 @@ namespace LibGit2Sharp.Tests
             string path = SandboxStandardTestRepo();
             using (var repo = new Repository(path))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
                 const string filename = "new_tracked_file.txt";
                 IndexEntry blob = repo.Index[filename];
 
@@ -112,7 +112,7 @@ namespace LibGit2Sharp.Tests
             string path = SandboxStandardTestRepo();
             using (var repo = new Repository(path))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
                 const string filename = "new_tracked_file.txt";
                 Assert.NotNull(repo.Index[filename]);
 
@@ -210,7 +210,7 @@ namespace LibGit2Sharp.Tests
             string path = SandboxStandardTestRepo();
             using (var repo = new Repository(path))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
 
                 string file = Path.Combine("Project", "a_file.txt");
 
@@ -282,7 +282,7 @@ namespace LibGit2Sharp.Tests
         {
             using (var repo = new Repository(SandboxStandardTestRepo()))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
 
                 repo.Stage(relativePath);
 
@@ -295,7 +295,7 @@ namespace LibGit2Sharp.Tests
         {
             using (var repo = new Repository(SandboxStandardTestRepo()))
             {
-                int count = repo.Index.Count;
+                long count = repo.Index.Count;
 
                 repo.Stage(new string[] { "*", "u*" });
